+++ choonchernlim-archetype-webapp-webapp/choonchernlim-archetype-webapp-webapp-war/src/main/frontend/src/js/app.js
@@ -2,15 +2,16 @@
 
 // the only way to get Bootstrap working is to make jQuery global... couldn't get
 // browserify-shim to work with Bootstrap and jQuery.
-var $ = global.jQuery = require( 'jquery' );
+import $ from 'jquery';
+global.jQuery = $;
 
-require( 'bootstrap' );
+import bootstrap from 'bootstrap';
 
 // getting scroll-spy content to align properly on menu click
-var offset = 100;
+const offset = 100;
 $( '#home-left-navbar' ).find( 'li a' ).click( function ( e ) {
-    var $a = $( this );
-    var $li = $( this ).parent();
+    const $a = $( this );
+    const $li = $( this ).parent();
 
     e.preventDefault();